# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VizTR Super Admin API",
    "version": "1.0.0",
    "description": "Super Admin Dashboard API for VizTR Architecture Visualization Platform",
    "contact": {
      "name": "VizTR Team",
      "email": "api@viztr.studio"
    }
  },
  "servers": [
    {
      "url": "https://api.viztr.studio",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/admin/users": {
      "get": {
        "summary": "List users",
        "description": "Get paginated list of admin users with filtering",
        "parameters": [
          {
            "schema": {
              "type": "object",
              "properties": {
                "search": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "super_admin",
                    "admin",
                    "user",
                    "client"
                  ]
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "active",
                    "invited",
                    "suspended",
                    "inactive"
                  ]
                },
                "department": {
                  "type": "string"
                },
                "page": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "default": 1
                },
                "pageSize": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 100,
                  "default": 20
                },
                "sortBy": {
                  "type": "string"
                },
                "sortOrder": {
                  "type": "string",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                }
              }
            },
            "required": false,
            "in": "query",
            "name": "UserFilters"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "email": {
                            "type": "string",
                            "format": "email"
                          },
                          "role": {
                            "type": "string",
                            "enum": [
                              "super_admin",
                              "admin",
                              "user",
                              "client"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "invited",
                              "suspended",
                              "inactive"
                            ]
                          },
                          "avatar": {
                            "type": "string",
                            "format": "uri"
                          },
                          "department": {
                            "type": "string"
                          },
                          "assignedProjectsCount": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "lastLogin": {
                            "type": "string"
                          },
                          "twoFactorEnabled": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "permissionsOverride": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "phone": {
                            "type": "string"
                          },
                          "company": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "email",
                          "role",
                          "status",
                          "avatar",
                          "department",
                          "assignedProjectsCount",
                          "lastLogin",
                          "twoFactorEnabled",
                          "createdAt"
                        ]
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "timestamp": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "summary": "Create user",
        "description": "Create a new admin user (super_admin required for admin roles)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "super_admin",
                      "admin",
                      "user",
                      "client"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "invited",
                      "suspended",
                      "inactive"
                    ]
                  },
                  "avatar": {
                    "type": "string",
                    "format": "uri"
                  },
                  "department": {
                    "type": "string"
                  },
                  "assignedProjectsCount": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "lastLogin": {
                    "type": "string"
                  },
                  "twoFactorEnabled": {
                    "type": "boolean"
                  },
                  "permissionsOverride": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "phone": {
                    "type": "string"
                  },
                  "company": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "role",
                  "status",
                  "avatar",
                  "department",
                  "assignedProjectsCount",
                  "lastLogin",
                  "twoFactorEnabled"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "super_admin",
                            "admin",
                            "user",
                            "client"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "active",
                            "invited",
                            "suspended",
                            "inactive"
                          ]
                        },
                        "avatar": {
                          "type": "string",
                          "format": "uri"
                        },
                        "department": {
                          "type": "string"
                        },
                        "assignedProjectsCount": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "lastLogin": {
                          "type": "string"
                        },
                        "twoFactorEnabled": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "permissionsOverride": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "phone": {
                          "type": "string"
                        },
                        "company": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "email",
                        "role",
                        "status",
                        "avatar",
                        "department",
                        "assignedProjectsCount",
                        "lastLogin",
                        "twoFactorEnabled",
                        "createdAt"
                      ]
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "timestamp": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
...
  },
...
}